FINAL VERSION
Hany's Revenge, An educational X86 Assembly 3D game!
The player is required to diffuse a bomb planted by Hany.
Luckily, Hany is so clumsy that he had left some clues in the office that has the code to diffuse the bomb.
the player has to search for the clues and diffuse the bomb before the time runs out!!